CV Intellegence Project/backend/requirements.txt
fastapi==0.110.0
uvicorn==0.30.1
transformers>=4.41.0
torch>=2.2.0
langchain>=0.2.3
langchain-community>=0.2.3
langchain-core>=0.2.3
pydantic<2
python-multipart>=0.0.9
pypdf>=4.2.0
python-dotenv>=1.0.1